Grounded secondary functional, primary improvements
The secondary has a 2.5 second charge time (scales with attack speed) and does 650%-1000% based on charge amount (damage scales linearly with charge). It has a 1 proc coeff, but if it's fully charged, it has 1.5 instead. It launches hit enemies in the direction you're aiming. It, like all launching attacks, has auto-aim. Given how freely you can aim it, I'm not sure if having the auto-aim on that is a good idea

Launch direction of primary is more consistent now. The downwards hit of hte primary combo, which normally launches downwards, now launches forwards if the target is on the ground

You're no longer able to use primary while boosting

Started some setup for Boost's hammer spin