Aerial secondary functional
Using the aerial secondary makes you start falling fast. It does the same damage as the grounded version, but instead of charging over time, it charges based on distance fallen similar to HEADST. Hitting an enemy while falling will launch them straight downwards. If you hit the ground instead, it will deal damage in an area around you. After the attack, you'll jump up very high in the air and get a small movement speed boost.

Made custom sized hitboxes for primary and secondary instead of using the default one. In general, the hitboxes are now a big bigger

Possible changes:

Make the big jump at the end only happen if you hold the skill down, so you don't always have to jump.
Hitting an enemy while falling means you don't do damage in an AOE, so it's bad to hit an enemy with it in certain cases. Changing this would require a decent bit of redoing things internally to work