New experiment with the secondary skill
Currently, the secondary skills starts charging if used on the ground or you go straight down and attack if used in the air. I realized that False Son has a very similar type of move with his club slam, but his functions a bit differently. His always starts charging the same way, but does a normal attack or goes down and attacks based on whether you're in the air or on the ground when you release the move. I'm not sure which one is better, so I made a version of Amy's secondary that functions more like False Son's move and made it an alt-skill.