Finished secondary skill and added placeholder skill icons
Officially moved to the alt secondary style from the previous commit.

Fixed the problem I mentioned in the previous commit about hitting an enemy with the aerial secondary not hitting in an AOE

Reduced secondary skill cooldown so it's easier to pogo off enemies over and over with it

Slightly buffed secondary damage

Added very rough versions of the skill icons. I'll probably just redo those same designs in higher quality for the final skill icons, keeping the same layout and whatever. I probably won't do any of the final skill icons until more of the animations for Amy are done, since I can use the posing of the animations to help make some of the icons