Started Multi-Lock and cleaned up some code stuff
If you're curious about the 'not colliding with enemies' issue I talked about, I looked into that, and found that it isn't a reliable option. I tried copying the logic that Mercenary's dashes use to pass through enemies, but it doesn't work on all enemies, like Solus Control Units. So I am going with a more experimental and complicated approach...
Multi-Lock uses similar logic to Huntress' glaive secondary. The idea is you'll disappear while the attack bounces between the targeted enemies. After the final hit, you'll teleport on top of the last enemy and reappear. I haven't done the teleporting part yet, but I do have the targeting and bouncing parts in.
After bouncing on an enemy, you attach a projectile that explodes after a short time, dealing damage in a small AOE. Not sure if the AOE damage is a good idea or if it should just damage the target. Since the plan is to make Super Multi-Lock do a big explosion, it's easy to slightly change the values on the normal projectile explosion to make the super projectile explosion
Multi-Lock is definitely broken in multiplayer. I'll think more about how to get it working in multiplayer once I finish the move

Created a new momentum passive icon. Not 100% satisfied with how it looks, so I'll probably change it a bit later. Once it's done, I will probably recolor it and replace Sonic's momentum passive icon with it too

Unified all files' namespaces