Character Select, vfx, scepter skill icon Amy now appears in the character select screen. No more stick figure
Added VFX for hitting enemies with the hammer. Added subtle sparkles around multi-lock hearts. Added new vfx for when secondary is fully charged

Added a buff icon for the buff from Scepter Multi-Lock

Added a new skill icon for Scepter Multi-Lock. Still need to come up with a new name for  the skill, since the scepter skills usually have cooler versions of the original name. "Multi-Lock" isn't a lot to work with